+++ app/lib/tool/backfill/backfill_new_fields.dart
@@ -3,6 +3,8 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:logging/logging.dart';
+import 'package:pub_dev/package/models.dart';
+import 'package:pub_dev/shared/datastore.dart';
 
 final _logger = Logger('backfill_new_fields');
 
@@ -12,5 +14,30 @@ final _logger = Logger('backfill_new_fields');
 /// CHANGELOG.md must be updated with the new fields, and the next
 /// release could remove the backfill from here.
 Future<void> backfillNewFields() async {
-  _logger.info('No new fields.');
+  _logger.info('Backfill admin deleted fields...');
+  await for (final e in dbService.query<Package>().run()) {
+    if (e.isAdminDeleted == null) {
+      await withRetryTransaction(dbService, (tx) async {
+        final p = await tx.lookupOrNull<Package>(e.key);
+        if (p == null) {
+          return;
+        }
+        p.isAdminDeleted ??= false;
+        tx.insert(p);
+      });
+    }
+  }
+
+  await for (final e in dbService.query<PackageVersion>().run()) {
+    if (e.isAdminDeleted == null) {
+      await withRetryTransaction(dbService, (tx) async {
+        final p = await tx.lookupOrNull<PackageVersion>(e.key);
+        if (p == null) {
+          return;
+        }
+        p.isAdminDeleted ??= false;
+        tx.insert(p);
+      });
+    }
+  }
 }