+++ CHANGELOG.md
@@ -4,6 +4,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
  * Bump runtimeVersion to `2025.05.21`.
  * Upgraded stable Flutter analysis SDK to `3.32.0`.
+ * Note: started to backfill `Package[Version]`'s `isAdminDeleted` field.
 
 ## `20250519t093200-all`